reconstructPoint int로 변경

+++ Assets/Source/TableClass/Runtime/InnerGradeTableData.cs
@@ -125,8 +125,8 @@ public class InnerGradeTableData
   public int transferGold { get { return _transferGold; } set { _transferGold = value; } }
   
   [SerializeField]
-  float _reconstructPoint;
-  public float reconstructPoint { get { return _reconstructPoint; } set { _reconstructPoint = value; } }
+  int _reconstructPoint;
+  public int reconstructPoint { get { return _reconstructPoint; } set { _reconstructPoint = value; } }
   
   [SerializeField]
   int _sellGold;
